packages/train-yolo/pyproject.toml
[project]
name = "labelr-train-yolo"
version = "0.1.0"
description = "A CLI to train Yolo object detection models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.15.1",
    "tqdm",
    "Pillow",
    "ultralytics==8.4.14",
    "albumentations",
    "labelr==0.14.0",
    "wandb==0.22.3",
    "torch==2.9.0",
    "Jinja2==3.1.6",
    # ONNX Runtime is necessary to export ONNX models
    "onnxruntime-gpu==1.23.2",
    "onnx==1.19.1",
    "onnxslim==0.1.72",
]

[tool.uv.sources]
torch = [
  { index = "pytorch", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
# Optional name for the index.
name = "pytorch"
# Required URL for the index.
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[[tool.uv.index]]
# Optional name for the index.
name = "nvidia"
# Required URL for the index.
url = "https://pypi.nvidia.com"
explicit = true
